requirements.txt
# --- AI Engine (CLIP + FAISS) ---
faiss-cpu
transformers
torch
Pillow

# --- API Server (FastAPI) ---
fastapi
uvicorn
python-multipart
